fix(review): address full-branch sweep findings for docs portal
Roll up Important findings from the pre-PR full-branch sweep
(`lfx-self-serve-code-reviewer`) of the 11 docs-portal commits.

GENERAL REVIEW
- docs-article click interceptor: scope to the `[innerHTML]` body
  container (`data-testid="docs-article-body"`) so framework-rendered
  anchors (breadcrumb/siblings via `[routerLink]`, search results via
  `DocsSearchComponent.activate()`) fall through to Angular's router
  instead of getting a redundant second `navigateByUrl()` call from
  the host-level listener.
- docs-article resolver: widen the type from `ResolveFn<DocsArticle>`
  to `ResolveFn<DocsArticle | RedirectCommand>` and drop the
  `as never` cast that was hiding the real return type.

REPO CONVENTIONS
- Move `DocsSearchHit` from a local export in
  `docs-search.service.ts` to `packages/shared/src/interfaces/`
  (component-organization §5: every interface that crosses files
  belongs in `@lfx-one/shared`).
- docs-article: extract the inline `toSignal(this.route.data...)`
  into a private `initArticle()` initializer
  (component-organization §3).
- docs-sidebar-nav: extract `isDocsActive` and `signInUrl`
  `toSignal` initializers into private `initIsDocsActive()` /
  `initSignInUrl()` (same rule). Also flip from `initialValue:` to
  `requireSync: true` since `startWith` already seeds the cold
  `router.events` source.
- lens-switcher: same `initIsDocsActive()` extraction +
  `requireSync: true`.
- sitemap.route: document the controller-service-route deviation
  inline — there's no business logic to abstract here (static
  build-artifact reader); split if/when it grows real logic.

NOTED ITEMS
- sanitize.mjs: drop `enforceHtmlBoundary: true` plus its
  misleading "Disable URL-rewriter's lower-casing" comment. The
  option strips control chars after `</html>` (per the
  sanitize-html README), which is irrelevant to fragment HTML
  emitted by `marked` — the comment described behavior that
  option doesn't have.
- build-manifest.mjs `computeSiblings`: exclude topic-landing
  articles from leaf siblings so "More in this topic" surfaces
  peer how-to articles only — the breadcrumb already exposes the
  parent landing.

Verification
- `yarn docs:check` (build + validate + coverage + sitemap +
  idempotence) — green
- `yarn lint:check` — green
- `yarn format:check` — green
- `yarn check-types` — green
- `yarn build` — green
